Sketch the desired model architecture
For starters, we'd want to use two convolutional layers.  It probably
makes sense to use max pooling, dropout, and ReLU activation on these.

Then we'll need to flatten if we want to pass to fully-connected layers.

Then, throw in a few fully-connected layers for good measure.

Since this is a regression problem (output a single decimal), our
penultimate layer needs to connect to a layer with a single node.